ARP: moved main for resolve, discover, and rdiscover to module examples
